Fix history state recall
Fixes going back in history not recalling previous state
(because photos failed to load).

Also don't show duplicates of duplicates.  Sometimes reddit doesn't
provide a full list, or they were cross-posted after we loaded the list.
So if we run into a "SELF" duplicate that alread exists, assume we are a
later duplciate, and don't show image.

Signed-off-by: Nathaniel Clark <Nathaniel.Clark@misrule.us>