r/news and better favicon support
This re-does the favicon logic, so that if the site isn't listed in
rp.favicons (should now be VERY limited) that is uses:
ORIGIN/favicon.ico and then falls back to the link icon if that fails to load
or if it's a 1x1 pixel.

This also adds support for reddit previews on items listed as T3.
So if the url isn't parsable, we load the preview and show that.
This adds support for reading r/news.

This also adds an extra link in comments box for Albums where we give a link
to story (parent.orig_url) as third icon.

Initial support for following crossposts, still need to acturatly reflect
and cleanup deuplicates list.

Signed-off-by: Nathaniel Clark <Nathaniel.Clark@misrule.us>