Fix late album load of first image
This makes div#albumNumberButtons an always present item,
so display and appending to it works.  Also fixes calls to
addNumberbutton(), which should always call with in an index.

Also allow reanimation if startAnimation() called while div is animating.
This allows a subsiquent call, for album load after image is displayed,
to be called once initial image is loaded.

Signed-off-by: Nathaniel Clark <Nathaniel.Clark@misrule.us>